# Conflicts:
#	packages/rs-drive-abci/tests/strategy_tests/upgrade_fork_tests.rs
#	packages/rs-platform-version/src/version/mod.rs

+++ packages/rs-drive/src/drive/credit_pools/epochs/credit_distribution_pools/get_epoch_fee_multiplier/mod.rs
@@ -28,7 +28,7 @@ impl Drive {
         epoch_tree: &Epoch,
         transaction: TransactionArg,
         platform_version: &PlatformVersion,
-    ) -> Result<f64, Error> {
+    ) -> Result<u64, Error> {
         match platform_version
             .drive
             .methods
